docker_compose_apps/project_docker_compose/backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
redis
PyJWT
Werkzeug
prometheus-flask-exporter
gunicorn
flask-cors